requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.4
django-appconf==1.0.2
django-classy-tags==0.8.0
django-easy-maps==0.9.3
django-embed-video==1.1.2
django-friendship==1.5.0
django-geoposition==0.3.0
django-google-maps==0.10.1
django-leaflet==0.23.0
django-map-widgets==0.1.9
django-simple-friends==1.1.1
geopy==1.12.0
googlemaps==2.5.1
gunicorn==19.7.1
idna==2.6
numpy==1.14.3
Pillow==5.1.0
psycopg2==2.7.4
pygal==2.4.0
pygame==1.9.3
pytz==2018.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
